e.player_answers fix (#556)
The `Array.prototype.includes()` method on Arrays uses the SameValueZero algorithm, which is basically strict equality except that it treats `NaN` as equivalent.

Since some CYOA mods use `e.player_answers.includes(answer)`, this will cause some bugs if the answer pushed is a String, and the answers are usually expected to be Number literals.

This fixes it by just checking whether t is a String and numT is NaN and if it is, it uses just t. Otherwise, it uses the coerced version of T. Since this is a ternary, it doesn't take into account other non-primitive values like Objects. But who even uses Objects for PK generation?